product/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
comm==0.1.2
DateTime==5.0
debugpy==1.6.6
decorator==5.1.1
django==4.0.7
celery==5.2.7
flower==1.2.0
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.9.0
django_debug_toolbar==3.8.1
django-environ==0.9.0
django-filter==22.1
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
ipykernel==6.21.1
ipython==8.9.0
jedi==0.18.2
jupyter_client==8.0.2
jupyter_core==5.2.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.3
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7.1
pyzmq==25.0.0
redis==4.4.2
setuptools==59.6.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
wcwidth==0.2.6
zope.interface==5.5.2